Refactored code and created components
