+++ src/components/AddToDatabase.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import AddItemsForm from './AddItemsForm';
+
+function AddToDatabase() {
+  return (
+    <div className="App">
+      <AddItemsForm />
+    </div>
+  );
+}
+
+export default AddToDatabase;
